Move exponential transforms into nested module
